+++ openwrt/internal/network/device.go
@@ -7,7 +7,9 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework-validators/int64validator"
 	"github.com/hashicorp/terraform-plugin-framework-validators/setvalidator"
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/path"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/joneshf/terraform-provider-openwrt/openwrt/internal/lucirpcglue"
@@ -190,6 +192,26 @@ var (
 		},
 	}
 )
+
+func NewDeviceDataSource() datasource.DataSource {
+	return lucirpcglue.NewDataSource(
+		deviceModelGetId,
+		deviceSchemaAttributes,
+		deviceSchemaDescription,
+		deviceUCIConfig,
+		deviceUCIType,
+	)
+}
+
+func NewDeviceResource() resource.Resource {
+	return lucirpcglue.NewResource(
+		deviceModelGetId,
+		deviceSchemaAttributes,
+		deviceSchemaDescription,
+		deviceUCIConfig,
+		deviceUCIType,
+	)
+}
 
 type deviceModel struct {
 	BridgePorts        types.Set    `tfsdk:"ports"`
