--- openwrt/internal/network/device_data_source.go
@@ -1,16 +1,0 @@
-package network
-
-import (
-	"github.com/hashicorp/terraform-plugin-framework/datasource"
-	"github.com/joneshf/terraform-provider-openwrt/openwrt/internal/lucirpcglue"
-)
-
-func NewDeviceDataSource() datasource.DataSource {
-	return lucirpcglue.NewDataSource(
-		deviceModelGetId,
-		deviceSchemaAttributes,
-		deviceSchemaDescription,
-		deviceUCIConfig,
-		deviceUCIType,
-	)
-}
