--- openwrt/internal/system/system_resource.go
@@ -1,16 +1,0 @@
-package system
-
-import (
-	"github.com/hashicorp/terraform-plugin-framework/resource"
-	"github.com/joneshf/terraform-provider-openwrt/openwrt/internal/lucirpcglue"
-)
-
-func NewSystemResource() resource.Resource {
-	return lucirpcglue.NewResource(
-		systemModelGetId,
-		systemSchemaAttributes,
-		systemSchemaDescription,
-		systemUCIConfig,
-		systemUCIType,
-	)
-}
